+++ tbot/loader.py
@@ -128,6 +128,13 @@ The testcase was defined in:
                         func._tbot_files = [f]
                         testcases[name] = func
         except:  # noqa: E722
-            print(c("Warning").yellow.bold + f": Failed to load {f}", file=sys.stderr)
+            import textwrap
+            import traceback
+
+            trace = textwrap.indent(traceback.format_exc(), "    ")
+            print(
+                c("Warning").yellow.bold + f": Failed to load {f}:\n{trace}",
+                file=sys.stderr,
+            )
 
     return testcases